+++ src/MetaheuristicsPlatform/Algorithms/Multimodal/DynamicNicheSharingGa/DynamicNicheSharingGaOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.Multimodal.DynamicNicheSharingGa;
+
+public static class DynamicNicheSharingGaOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Miller & Shaw",
+            1996,
+            "Genetic algorithms with dynamic niche sharing for multimodal function optimization",
+            "Proceedings of the IEEE International Conference on Evolutionary Computation, 786-791",
+            "10.1109/ICEC.1996.542701");
+}